~Grace helped with the logic
Updated the project passing cart around in the session instead of having to constantly update it on the server, would be nice if a user logs out or exits the app for the cart to be updated before they do so. If a user doesn't have a cart it creates one for them when they go to add a game. If a game doesn't exist in the cart it adds it to the cart, along with the quantity being updated correctly. GamesForCart model was edited to use Game models instead of the m_item1 model that was causing issues.