Grace made the adapter for the carts listview and had the code to make the listview work, there was a small error that I fixed to allow it to actually display correctly. Hooked the listview items in the cart up to the cartdetailsfragment that already existed.
Cart is now updated when user closes the app, so cart is returned correctly when they come back. Removed getCartId from session info since we didnt actually need it. Need to go back and rethink the design for the carts page since the listview pushes the ready to checkout button off the view. Going to work on cleaning up what we have already so not so much code is reused and front end stuff to make the app look spiffy.